requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
streamlit==1.36.0
python-multipart==0.0.9
python-docx==1.1.0
PyMuPDF==1.24.5 # This is 'fitz'
pandas==2.2.2
openpyxl==3.1.3 # Needed for reading .xlsx files with pandas
langchain==0.2.6
langchain-community==0.2.6
langchain-ollama==0.2.6
faiss-cpu==1.8.0 # Or faiss-gpu if you have a GPU and want to use it
pydantic-settings==2.3.4
requests